+++ pkg/cmd/templates/templater.go
@@ -7,7 +7,7 @@ import (
 	"text/template"
 	"unicode"
 
-	"github.com/openshift/origin/pkg/cmd/util/term"
+	"github.com/openshift/oc/pkg/helpers/term"
 	ktemplates "k8s.io/kubernetes/pkg/kubectl/util/templates"
 
 	"github.com/spf13/cobra"
