+++ pkg/oc/cli/admin/prune/builds/resolvers_test.go
@@ -10,7 +10,7 @@ import (
 	"k8s.io/apimachinery/pkg/util/sets"
 
 	buildv1 "github.com/openshift/api/build/v1"
-	"github.com/openshift/origin/pkg/oc/lib/buildapihelpers"
+	buildapihelpers "github.com/openshift/oc/pkg/helpers/build"
 )
 
 type mockResolver struct {
