+++ pkg/oc/cli/logout/logout.go
@@ -16,8 +16,8 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/util/templates"
 
 	oauthv1client "github.com/openshift/client-go/oauth/clientset/versioned/typed/oauth/v1"
-	kubeconfiglib "github.com/openshift/origin/pkg/oc/lib/kubeconfig"
-	"github.com/openshift/origin/pkg/oc/util/project"
+	kubeconfiglib "github.com/openshift/oc/pkg/helpers/kubeconfig"
+	"github.com/openshift/oc/pkg/helpers/project"
 )
 
 type LogoutOptions struct {
