+++ pkg/oc/cli/projects/projects.go
@@ -17,10 +17,10 @@ import (
 
 	projectv1 "github.com/openshift/api/project/v1"
 	projectv1client "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
+	cliconfig "github.com/openshift/oc/pkg/helpers/kubeconfig"
 	oapi "github.com/openshift/origin/pkg/api"
 	clientcfg "github.com/openshift/origin/pkg/client/config"
 	ocproject "github.com/openshift/origin/pkg/oc/cli/project"
-	cliconfig "github.com/openshift/origin/pkg/oc/lib/kubeconfig"
 )
 
 type ProjectsOptions struct {
