+++ pkg/oc/cli/requestproject/request_project.go
@@ -13,8 +13,8 @@ import (
 
 	projectv1 "github.com/openshift/api/project/v1"
 	projectv1client "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
+	cliconfig "github.com/openshift/oc/pkg/helpers/kubeconfig"
 	ocproject "github.com/openshift/origin/pkg/oc/cli/project"
-	cliconfig "github.com/openshift/origin/pkg/oc/lib/kubeconfig"
 )
 
 // RequestProjectOptions contains all the options for running the RequestProject cli command.
