+++ pkg/oc/cli/status/status.go
@@ -18,8 +18,8 @@ import (
 	imagev1client "github.com/openshift/client-go/image/clientset/versioned/typed/image/v1"
 	projectv1client "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
 	routev1client "github.com/openshift/client-go/route/clientset/versioned/typed/route/v1"
+	loginutil "github.com/openshift/oc/pkg/helpers/project"
 	"github.com/openshift/origin/pkg/oc/lib/describe"
-	loginutil "github.com/openshift/origin/pkg/oc/util/project"
 	dotutil "github.com/openshift/origin/pkg/util/dot"
 )
