+++ pkg/oc/lib/describe/describer.go
@@ -46,6 +46,8 @@ import (
 	buildv1clienttyped "github.com/openshift/client-go/build/clientset/versioned/typed/build/v1"
 	onetworktypedclient "github.com/openshift/client-go/network/clientset/versioned/typed/network/v1"
 	quotaclient "github.com/openshift/client-go/quota/clientset/versioned/typed/quota/v1"
+	ocbuildapihelpers "github.com/openshift/oc/pkg/helpers/build"
+	routedisplayhelpers "github.com/openshift/oc/pkg/helpers/route"
 	oapi "github.com/openshift/origin/pkg/api"
 	"github.com/openshift/origin/pkg/api/legacy"
 	authorizationapi "github.com/openshift/origin/pkg/authorization/apis/authorization"
@@ -56,8 +58,6 @@ import (
 	imageapi "github.com/openshift/origin/pkg/image/apis/image"
 	imageclient "github.com/openshift/origin/pkg/image/generated/internalclientset/typed/image/internalversion"
 	oauthclient "github.com/openshift/origin/pkg/oauth/generated/internalclientset/typed/oauth/internalversion"
-	ocbuildapihelpers "github.com/openshift/origin/pkg/oc/lib/buildapihelpers"
-	"github.com/openshift/origin/pkg/oc/lib/routedisplayhelpers"
 	projectapi "github.com/openshift/origin/pkg/project/apis/project"
 	projectclient "github.com/openshift/origin/pkg/project/generated/internalclientset/typed/project/internalversion"
 	quotaconvert "github.com/openshift/origin/pkg/quota/apis/quota"
