+++ pkg/oc/lib/describe/projectstatus.go
@@ -36,10 +36,11 @@ import (
 	imagev1client "github.com/openshift/client-go/image/clientset/versioned/typed/image/v1"
 	projectv1client "github.com/openshift/client-go/project/clientset/versioned/typed/project/v1"
 	routev1client "github.com/openshift/client-go/route/clientset/versioned/typed/route/v1"
+	loginerrors "github.com/openshift/oc/pkg/helpers/errors"
+	routedisplayhelpers "github.com/openshift/oc/pkg/helpers/route"
 	oapi "github.com/openshift/origin/pkg/api"
 	appsutil "github.com/openshift/origin/pkg/apps/util"
 	buildapi "github.com/openshift/origin/pkg/build/apis/build"
-	loginerrors "github.com/openshift/origin/pkg/oc/lib/errors"
 	appsedges "github.com/openshift/origin/pkg/oc/lib/graph/appsgraph"
 	appsanalysis "github.com/openshift/origin/pkg/oc/lib/graph/appsgraph/analysis"
 	appsgraph "github.com/openshift/origin/pkg/oc/lib/graph/appsgraph/nodes"
@@ -56,7 +57,6 @@ import (
 	routeedges "github.com/openshift/origin/pkg/oc/lib/graph/routegraph"
 	routeanalysis "github.com/openshift/origin/pkg/oc/lib/graph/routegraph/analysis"
 	routegraph "github.com/openshift/origin/pkg/oc/lib/graph/routegraph/nodes"
-	"github.com/openshift/origin/pkg/oc/lib/routedisplayhelpers"
 	"github.com/openshift/origin/pkg/util/errors"
 	"github.com/openshift/origin/pkg/util/parallel"
 )
