+++ pkg/oc/originpolymorphichelpers/logsforobject.go
@@ -16,9 +16,9 @@ import (
 	buildv1 "github.com/openshift/api/build/v1"
 	appsv1client "github.com/openshift/client-go/apps/clientset/versioned/typed/apps/v1"
 	buildv1client "github.com/openshift/client-go/build/clientset/versioned/typed/build/v1"
+	ocbuildapihelpers "github.com/openshift/oc/pkg/helpers/build"
 	appsmanualclientv1 "github.com/openshift/origin/pkg/apps/client/v1"
 	buildmanualclientv1 "github.com/openshift/origin/pkg/build/client/v1"
-	ocbuildapihelpers "github.com/openshift/origin/pkg/oc/lib/buildapihelpers"
 )
 
 func NewLogsForObjectFn(delegate polymorphichelpers.LogsForObjectFunc) polymorphichelpers.LogsForObjectFunc {
