+++ staging/src/github.com/openshift/oc/pkg/helpers/doc.go
@@ -1,4 +1,4 @@
-// lib contains the logic of commands used by oc.  It should never depend on any CLI constructs like commands or flags
+// helpers contains the logic of commands used by oc.  It should never depend on any CLI constructs like commands or flags
 // it should be entirely focused on runtime need.
 // We're starting with a coarse chop
-package lib
+package helpers
