+++ staging/src/github.com/openshift/oc/pkg/helpers/route/status.go
@@ -1,4 +1,4 @@
-package routedisplayhelpers
+package route
 
 import (
 	corev1 "k8s.io/api/core/v1"
